# output: Improve layout path construction