The existing layout path construction was inefficient and rather crude.
Create a simple function to efficiently build the path string.

Also, fix the benchmark function to create the layout handler inside the loop.
Otherwise, the go benchmark framework would run the benchmark once to
test it's speed in order to decide the value of b.N.  Afterward, the
layout handler had already cached all the work.  I spent way too much
time scratching my head at why the benchmarks didn't improve with the
new path construction function.
```
name      old time/op    new time/op    delta
Layout-2    77.7µs ± 8%    14.0µs ±14%  -81.98%  (p=0.008 n=5+5)

name      old alloc/op   new alloc/op   delta
Layout-2    41.9kB ± 0%     6.3kB ± 0%  -84.88%  (p=0.016 n=5+4)

name      old allocs/op  new allocs/op  delta
Layout-2       487 ± 0%       112 ± 0%  -77.00%  (p=0.008 n=5+5)
```