@bep 2020-10-13
>Afterward, the layout handler had already cached all the work.

One could argue that that is the correct behaviour for this benchmark. Caches is why most fast software is fast, so they need to be in the benchmarks to get real numbers. It was certainly intended. I suggest you duplicate the Benchmark and name it something else.