@bep 2020-10-13
Looks good. Can you squash + reword the commit message. I'm not a fan of commit messages that takes a stab at the existing implementation for no particular great reason. A method that takes 77 microseconds to run where the result is cached is plenty fast enough. But I agree that your implementation is cleaner.